Committer: Chris Warburton <chris@chris-portable.(none)>  On branch master  Changes to be committed: 	modified:   pubsub_reader.py 	modified:   pubsub_writer.py 	modified:   pubsubclient.py 	modified:   pubsubclient.pyc
Started reworking the reader test app to use the current API. It can find nodes to subscribe
to, but subscribing is just a stub at the moment.

Also disabled importing Webkit until I can get it installed on Debian. The browser display is
useless until we can subscribe to and receive updates from nodes anyway.